Updated the adImpression message to include event details along with user and campaign data.
