+++ go.mod
@@ -5,6 +5,7 @@ go 1.22.4
 require github.com/segmentio/kafka-go v0.4.47
 
 require (
+	github.com/google/uuid v1.6.0 // indirect
 	github.com/klauspost/compress v1.15.9 // indirect
 	github.com/pierrec/lz4/v4 v4.1.15 // indirect
 )
